move Blocks components to Blocks directory
